refactor python code in separate modules
